+++ app/uk/gov/hmrc/tradergoodsprofiles/models/requests/router/RouterUpdateRecordRequest.scala
@@ -31,7 +31,7 @@ case class RouterUpdateRecordRequest(
   countryOfOrigin: Option[String] = None,
   category: Option[Int] = None,
   assessments: Option[Seq[Assessment]] = None,
-  supplementaryUnit: Option[Int] = None,
+  supplementaryUnit: Option[BigDecimal] = None,
   measurementUnit: Option[String] = None,
   comcodeEffectiveFromDate: Option[Instant] = None,
   comcodeEffectiveToDate: Option[Instant] = None
